+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Put finalizer on `cluster-values` configmap to keep base domain value until deletion finished.
+
 ## [0.11.2] - 2023-03-15
 
 ### Fixed